import inline js into head component
