+++ ui/package.json
@@ -59,6 +59,7 @@
         "node-sass": "^4.5.3",
         "preact-svg-loader": "^0.2.0",
         "prettier": "^1.5.3",
+        "raw-loader": "^0.5.1",
         "sass-loader": "^6.0.6",
         "stylelint-config-standard": "^17.0.0",
         "stylelint-order": "^0.6.0",
